fix: wait for pcmanfm process and socket
The current logic for waiting for pcmanfm to be ready seems to trigger
another race-condition on 64-bit systems, where the process is already
running but not fully initialized yet.

With this change, the wallpaper service will not be launched until the
process was found and a valid pcmanfm socket has been detected. The
socket path has been taken from the pcmanfm source [1].

[1] https://github.com/lxde/pcmanfm/blob/2c0a6b80dab1cd97f0a408ebbfa2f3e738f80648/src/single-inst.c#L405-L429